clients/rpi_client/requirements.txt
PyAudio
ffmpeg-python
scipy
websocket-client
python-dotenv
requests
alsa-utils
